refactor: migrate guardrails, preflight, plugins, telemetry, precision, profiles to drift-engine
(cherry picked from commit ccec60fc9fcf19272e5debdffe68af1c7cca9992)
